Move constants to transferconfig
