+++ awscli/customizations/s3/s3handler.py
@@ -16,8 +16,8 @@ import math
 import os
 import sys
 
-from awscli.customizations.s3.constants import MULTI_THRESHOLD, CHUNKSIZE, \
-    NUM_THREADS, MAX_QUEUE_SIZE
+from awscli.customizations.s3.transferconfig import MULTI_THRESHOLD, \
+    CHUNKSIZE, NUM_THREADS, MAX_QUEUE_SIZE
 from awscli.customizations.s3.utils import find_chunksize, \
     operate, find_bucket_key, relative_path, PrintTask, create_warning
 from awscli.customizations.s3.executor import Executor
